refactor(GLBuffer): refactor GLBuffer management into separate files for better organization
