+++ src/plugin/opengl/xmake.lua
@@ -28,6 +28,7 @@ target("PluginOpenGL")
     add_includedirs("src/utils/Light", {public = true})
     add_includedirs("src/component/Model/", {public = true})
     add_includedirs("src/resource/ShaderManager/", {public = true})
+    add_includedirs("src/resource/GLBufferManager/", {public = true})
     add_includedirs("src/resource/MaterialCache/", {public = true})
     add_includedirs("src/resource/Camera/", {public = true})
     add_includedirs("src/resource/GLBuffer/", {public = true})
